+++ plot/bars.go
@@ -19,6 +19,8 @@ import (
 // Creates a bar per column of the observer.
 type Bars struct {
 	Observer observer.Row // Observer providing a data series for bars.
+	YLim     [2]float64   // Y axis limits. Optional, default auto.
+	Labels   Labels       // Labels for plot and axes. Optional.
 
 	headers []string
 	series  plotter.Values
@@ -57,13 +59,12 @@ func (b *Bars) Draw(w *ecs.World, win *pixelgl.Window) {
 	c := vgimg.New(vg.Points(width*b.scale)-10, vg.Points(height*b.scale)-10)
 
 	p := plot.New()
+	setLabels(p, b.Labels)
 
-	p.X.Tick.Label.Font.Size = 12
-	p.X.Tick.Label.Font.Variant = "Mono"
-
-	p.Y.Tick.Label.Font.Size = 12
-	p.Y.Tick.Label.Font.Variant = "Mono"
-	p.Y.Tick.Marker = paddedTicks{}
+	if b.YLim[0] != 0 || b.YLim[1] != 0 {
+		p.Y.Min = b.YLim[0]
+		p.Y.Max = b.YLim[1]
+	}
 
 	bw := 0.5 * (width - 50) / float64(len(b.series))
 	bars, err := plotter.NewBarChart(b.series, vg.Points(bw))